修改eTPSS数据脱敏api

+++ src/main/java/com/xd/hufei/dto/other/TableColumn.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.ToString;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @ToString
 public class TableColumn {
     private String columnName;
     private String dataType;
-    private Integer characterMaximumLength;
+    private BigInteger characterMaximumLength;
 }
